+++ workspace/solutions/rust/src/s0/mod.rs
@@ -1,2 +1,3 @@
 pub mod s0;
+pub mod s4;
 pub mod s9;