+++ workspace/solutions/rust/src/s0/s4/s1/s5/mod.rs
@@ -1,0 +1,4 @@
+pub mod s_0415_00;
+pub mod s_0415_test;
+
+pub type TSolution = fn(String, String) -> String;